fixed a bug where we subtly change the classifier id between releases
Summary:
since we extended the length of the classifiers array the hash function now
appends a series of '.'s to the end of each classifier.  This didn't happen
before since we used the right number of elements.

This fix causes the hash function to mirror the full length of the current
classifier without appending "empty" elements.

Test Plan: manual + shutterstock dev environment

Reviewers: carlos, austin, wsong, ankur

Reviewed By: ankur

Differential Revision: http://grizzly.memsql.com/D8362